# openapi.yml
openapi: 3.1.0
info:
  description: The same-origin contract consumed by the BeyondPilot web application.
  title: BeyondPilot API
  version: 1.0.0
servers:
- description: Same origin as the web application
  url: /
...
paths:
...
  /api/organization/invitations/{id}/accept:
    post:
      operationId: acceptOrganizationInvitation
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The caller belongs to the organization.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No open invitation to the caller has this identifier.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller already belongs to an organization.
      security:
      - session: []
      summary: Accept an invitation to the caller's address
      tags:
      - Organizations
  /api/organization/invitations/{id}/decline:
    post:
      operationId: declineOrganizationInvitation
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The invitation is closed.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No open invitation to the caller has this identifier.
      security:
      - session: []
      summary: Decline an invitation to the caller's address
      tags:
      - Organizations
  /api/organization/join-request/withdraw:
    post:
      operationId: withdrawJoinRequest
      responses:
        "204":
          description: The caller waits on no request.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
      security:
      - session: []
      summary: Withdraw the request the caller waits on
      tags:
      - Organizations
...
  /api/organization/organizations:
    get:
      operationId: searchOrganizations
      parameters:
      - description: "A part of the name, ignoring case."
        in: query
        name: q
        required: false
        schema:
          type: string
          maxLength: 100
          minLength: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationSearch"
          description: "Ten organizations at most, by name; none for a text shorter\
            \ than two characters."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The text is too long.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
      security:
      - session: []
      summary: The approved organizations whose name contains a text
      tags:
      - Organizations
    post:
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganization"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: "The organization, waiting for GenAI Fund's review."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller already belongs to an organization, or waits on\
            \ a request to join one."
      security:
      - session: []
      summary: Create an organization the caller owns
      tags:
      - Organizations
  /api/organization/organizations/{id}/join:
    post:
      operationId: joinOrganization
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinOrganization"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinOutcome"
          description: "What asking did: joined, owner, or a request that waits."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The message is too long.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such approved organization.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller already belongs to an organization, or waits on\
            \ a request."
      security:
      - session: []
      summary: Ask to get into an organization
      tags:
      - Organizations
  /api/organization/organizations/{slug}:
    get:
      operationId: getOrganization
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrganization"
          description: The organization.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No approved organization has this address.
      summary: "The public page of an approved organization, read without a session"
      tags:
      - Organizations
...
components:
  schemas:
...
    CreateOrganization:
      type: object
      description: "A new organization. Its creator owns it, and it waits for GenAI\
        \ Fund's review."
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
          pattern: "[A-Z]{2}"
        description:
          type:
          - string
          - "null"
          maxLength: 2000
          minLength: 0
        name:
          type: string
          maxLength: 120
          minLength: 0
        roles:
          type: array
          description: One or both of `provider` and `enterprise`.
          items:
            type: string
            pattern: provider|enterprise
          maxItems: 2
          minItems: 1
        teamSize:
          type:
          - string
          - "null"
          pattern: just_me|2_9|10_49|50_99|100_499|500_999|1000_4999|5000_plus
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
          pattern: company|builder_team|independent_builder|other
        website:
          type:
          - string
          - "null"
          maxLength: 300
          minLength: 0
          pattern: "https?://[^\\s]+"
      required:
      - name
      - roles
      - type
...
    JoinOrganization:
      type: object
      description: What a person says with a request to join or to claim.
      properties:
        message:
          type:
          - string
          - "null"
          maxLength: 1000
          minLength: 0
    JoinOutcome:
      type: object
      description: What asking to get into an organization did.
      properties:
        outcome:
          type: string
          description: "`joined`: the caller is a member. `owner`: nobody owned it\
            \ and the caller's address is on its domain, so they own it. `requested`:\
            \ its owners, or GenAI Fund when nobody owns it, decide."
          enum:
          - joined
          - owner
          - requested
      required:
      - outcome
...
    Organization:
      type: object
      description: "An organization as the people who belong to it, and operators,\
        \ see it."
      properties:
        autoJoin:
          type: boolean
          description: Whether an address on the domain joins at once; otherwise it
            asks the owners.
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        createdAt:
          type: string
          format: date-time
        decisionMessage:
          type:
          - string
          - "null"
          description: What the operator wrote to the owners with the refusal.
        decisionReason:
          type:
          - string
          - "null"
          description: Why it was last refused.
          enum:
          - duplicate
          - not_a_real_organization
          - incomplete
          - out_of_scope
          - other
        description:
          type:
          - string
          - "null"
        emailDomain:
          type:
          - string
          - "null"
          description: The domain whose addresses may join; null when it was made
            from a public mail address.
        id:
          type: string
          format: uuid
        name:
          type: string
        roles:
          type: array
          description: "What it does here: `provider` lists AI solutions, `enterprise`\
            \ posts use cases."
          items:
            type: string
        slug:
          type: string
        status:
          type: string
          description: GenAI Fund's review of the organization.
          enum:
          - pending
          - approved
          - rejected
        teamSize:
          type:
          - string
          - "null"
          description: Null is unknown.
          enum:
          - just_me
          - 2_9
          - 10_49
          - 50_99
          - 100_499
          - 500_999
          - 1000_4999
          - 5000_plus
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
        version:
          type: integer
          format: int64
          description: "Sent back with a save, which is refused when the organization\
            \ changed since."
        website:
          type:
          - string
          - "null"
      required:
      - autoJoin
      - createdAt
      - id
      - name
      - roles
      - slug
      - status
      - type
      - version
...
    OrganizationMatch:
      type: object
      description: "An organization a person may get into, and how."
      properties:
        country:
          type:
          - string
          - "null"
        emailDomain:
          type:
          - string
          - "null"
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
        way:
          type: string
          description: "What asking to get in does for this caller: `join` makes them\
            \ a member at once (owner, when nobody owns it yet), `request` asks its\
            \ owners, `claim` asks GenAI Fund to let them own it."
          enum:
          - join
          - request
          - claim
      required:
      - id
      - name
      - type
      - way
...
    OrganizationSearch:
      type: object
      description: "The approved organizations a search finds, by name; ten at most."
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMatch"
      required:
      - items
    Problem:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          description: Stable failure code for module failures and request validation;
            clients branch on it.
          maxLength: 63
          pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
        detail:
          type: string
          description: Safe human-readable explanation; never branch on it.
        errors:
          type: array
          description: "Request validation only: one entry per violation. A rejected\
            \ value is never echoed."
          items:
            type: object
            additionalProperties: false
            properties:
              code:
                type: string
                description: "Constraint code, for example `NOT_BLANK` or `SIZE`."
                pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
              detail:
                type: string
                description: Safe fallback text; clients translate by code.
              params:
                type: object
                additionalProperties: false
                description: Numeric bounds of the constraint; omitted when it has
                  none.
                properties:
                  max:
                    type: number
                  min:
                    type: number
              pointer:
                type: string
                description: "JSON Pointer to the offending member, for example `#/title`."
            required:
            - code
            - detail
            - pointer
        instance:
          type: string
          format: uri-reference
          description: Request path that produced the problem.
        requestId:
          type: string
          format: uuid
          description: Identifier of the request in the server logs and the `X-Request-Id`
            header.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        title:
          type: string
          description: Short summary of the problem kind; never branch on it.
        type:
          type: string
          format: uri
          description: '`urn:beyondpilot:failure:<code in kebab case>` when `code`
            is present; omitted means `about:blank`.'
      required:
      - requestId
      - status
      - title
...
    PublicOrganization:
      type: object
      description: An approved organization as anyone reads it.
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        description:
          type:
          - string
          - "null"
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
        website:
          type:
          - string
          - "null"
      required:
      - name
      - slug
      - type
...
  securitySchemes:
    session:
      description: "The session cookie set by signing in. Requests that change state\
        \ also send `X-BeyondPilot-CSRF: 1`."
      in: cookie
      name: BEYONDPILOT_SESSION
      type: apiKey